+++ R/geneticAlgorithm.R
@@ -452,6 +452,12 @@ GAfit <- function(data, p, M, model=c("GMVAR", "StMVAR", "G-StMVAR"), conditiona
     }
 
     # Take care of individuals that are not good enough + calculate the numbers redundant regimes
+    if(anyNA(logliks[i1,])) {
+      which_inds_na <- which(is.na(logliks[i1,]))
+      logliks[i1, which_inds_na] <- minval
+      warning(paste0("Removed NaNs from the logliks of inds ", which_inds_na, " in the GA round ", i1,
+                     " with the seed ", seed, ". To fix the issue, do a bug report (if the seed is not NULL)."))
+    }
     logliks[i1, which(logliks[i1,] < minval)] <- minval
     redundants[i1, which(logliks[i1,] <= minval)] <- M