+++ NEWS.md
@@ -163,3 +163,4 @@
 
 * Changed the default ncalls in fitGSMVAR to (M + 1)^5.
 * Fixed a bug that sometimes caused an error in the estimation algorithm because the random degrees of freedom parameters were too close to their strict lower bound 2.
+* Adjusted the tolerance when the estimation functions warns about near-non-stationary estimates or near-singular error term covariance matrices.