+++ NEWS.md
@@ -161,6 +161,7 @@
 
 # gmvarkit 2.0.2
 
+* We do not recycle the class "htest" for Wald and LR tests anymore because the print method was buggy and beyond the control of the author. Now Wald and LR tests use the class "hypotest" with a dedicated print method.
 * Changed the default ncalls in fitGSMVAR to (M + 1)^5.
-* Fixed a bug that sometimes caused an error in the estimation algorithm because the random degrees of freedom parameters were too close to their strict lower bound 2.
+* Fixed a bug that sometimes caused an error in the estimation algorithm because the random degrees of freedom parameters were too close to their strict lower bound 2. This change may have an effect on the estimation results of StMVAR and G-StMVAR models with specific seed.
 * Adjusted the tolerance when the estimation functions warns about near-non-stationary estimates or near-singular error term covariance matrices.